fix: Remove precision cap from antenna info
Refactor methods to format utils
